+++ pkg/instructions/block.go
@@ -1,0 +1,7 @@
+package instructions
+
+type Block interface {
+	RenderTerminal() string
+	RenderString() string
+	RenderMarkdown() string
+}
